towards-targets: Move mpi routines to separate file
