+++ devito/targets/common/__init__.py
@@ -1,5 +1,6 @@
 from .engine import *  # noqa
 from .blocking import *  # noqa
 from .openmp import *  # noqa
+from .mpi import *  # noqa
 from .rewriters import *  # noqa
 from .definitions import *  # noqa
